Enhance CMake configuration by adding ASIO library support and improving type safety in request/response handling

+++ src/server.cpp
@@ -8,7 +8,7 @@
 
 namespace network
 {
-    server::server(std::string_view host, unsigned short port, std::size_t concurrency_hint) : io_ctx(concurrency_hint), signals(io_ctx, SIGINT, SIGTERM, SIGQUIT), endpoint(asio::ip::make_address(host), port), acceptor(asio::make_strand(io_ctx))
+    server::server(std::string_view host, unsigned short port, std::size_t concurrency_hint) : io_ctx(static_cast<int>(concurrency_hint)), signals(io_ctx, SIGINT, SIGTERM, SIGQUIT), endpoint(asio::ip::make_address(host), port), acceptor(asio::make_strand(io_ctx))
     {
         threads.reserve(concurrency_hint);
         signals.async_wait([this](const std::error_code &ec, [[maybe_unused]] int signal)
